chore(yarn.lock): update dependencies for "@lambda-group/scylladb" and "@napi-rs/cli" packages
The "@lambda-group/scylladb" package has been updated to version 0.0.0-use.local and now has dependencies on "@napi-rs/cli" version ^2.16.3, ava version ^5.1.1, and tsx version ^3.13.0. The link type for this package is set to soft.

The "@napi-rs/cli" package has been updated to version 2.16.3. The link type for this package is set to hard.

The "semver" package has been updated to version 7.5.4.