fix(CI.yml): update OPENSSL_DIR and OPENSSL_STATIC environment variables for the Build job
The OPENSSL_DIR and OPENSSL_STATIC environment variables were updated in the Build job to use specific values for the 'windows-latest' host. The OPENSSL_DIR was set to 'C:\vcpkg\installed\x64-windows-static-md' and OPENSSL_STATIC was set to 'YES'. This change ensures that the correct OpenSSL configuration is used during the build process.