feat(scylla): upgrade Scylla driver and add UDT support
This commit upgrades the Scylla driver to version 0.13.1 and adds support for User Defined Types (UDTs). The `QueryParameter` struct has been updated to handle UDTs and the `QueryResult` struct now parses UDTs correctly. The `execute`, `query`, and `batch` methods in `ScyllaSession` have been updated to handle parameters of UDTs.

The `Uuid` struct has been updated to be cloneable and copyable.

Signed-off-by: Daniel Boll <danielboll.academico@gmail.com>