feat(types): add support for Decimal and Duration types
This commit introduces support for Decimal and Duration types in the
Scylla driver. It also updates the query parameter and result parsing
to handle these new types. The changes are reflected in the TypeScript
definitions and JavaScript exports as well.

BREAKING CHANGE: The query parameter and result types have been updated
to accommodate the new Decimal and Duration types. This may require
changes in the calling code if it was relying on the previous types.

Signed-off-by: Daniel Boll <danielboll.academico@gmail.com>