build: add MUSL target setup for CI builds (#46)
This pull request includes an update to the CI configuration to support
MUSL targets. The most important change is the addition of a step to set
up MUSL target and tools in the CI workflow.

CI configuration updates:

*
[`.github/workflows/CI.yml`](diffhunk://#diff-3ab46ee209a127470fce3c2cf106b1a1dbadbb929a4b5b13656a4bc4ce19c0b8R113-R118):
Added a step to set up MUSL target and tools, including updating the
package list, installing `musl-tools`, and adding the MUSL target using
`rustup` if the target contains 'musl'.